requirements.txt
discord.py[voice]==2.4.0
